feat: get,create,edit,delete cabin methods what used with react query converting into custom hooks.
