+++ src/features/cabins/CabinRow.jsx
@@ -1,10 +1,9 @@
+//@ts-nocheck
 import styled from "styled-components";
 import { formatCurrency } from "../../utils/helpers";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { deleteCabin } from "../../services/apiCabins";
-import toast from "react-hot-toast";
 import { useState } from "react";
 import CreateCabinForm from "./CreateCabinForm";
+import { useDeleteCabin } from "./useDeleteCabin";
 
 const TableRow = styled.div`
   display: grid;
@@ -56,17 +55,8 @@ export default function CabinRow({ cabin }) {
     image,
   } = cabin;
 
-  const queryClient = useQueryClient();
-  const { isLoading: isDeleting, mutate } = useMutation({
-    mutationFn: deleteCabin,
-    onSuccess: () => {
-      toast.success("Cabin succesfully deleted.");
-      queryClient.invalidateQueries({
-        queryKey: ["cabin"],
-      });
-    },
-    onError: (err) => toast.error(err.message),
-  });
+  const { isDeleting, deleteCabin } = useDeleteCabin();
+
   return (
     <>
       <TableRow role="table">
@@ -86,7 +76,7 @@ export default function CabinRow({ cabin }) {
             onClick={() => {
               const answer = confirm("Are u sure");
               if (answer) {
-                mutate(cabinId);
+                deleteCabin(cabinId);
               }
             }}
             disabled={isDeleting}
