+++ src/features/cabins/useCabins.jsx
@@ -1,0 +1,14 @@
+import { useQuery } from "@tanstack/react-query";
+import { getCabins } from "../../services/apiCabins";
+
+export function useCabins() {
+  const {
+    isLoading,
+    data: cabins,
+    error,
+  } = useQuery({
+    queryKey: ["cabin"],
+    queryFn: getCabins,
+  });
+  return { isLoading, cabins, error };
+}
