+++ src/features/cabins/useDeleteCabin.jsx
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import toast from "react-hot-toast";
+import { deleteCabin as deleteCabinApi } from "../../services/apiCabins";
+
+export function useDeleteCabin() {
+  const queryClient = useQueryClient();
+  const { isLoading: isDeleting, mutate: deleteCabin } = useMutation({
+    mutationFn: deleteCabinApi,
+    onSuccess: () => {
+      toast.success("Cabin succesfully deleted.");
+      queryClient.invalidateQueries({
+        queryKey: ["cabin"],
+      });
+    },
+    //@ts-ignore
+    onError: (err) => toast.error(err.message),
+  });
+
+  return { isDeleting, deleteCabin };
+}
