+++ src/pages/Cabins.jsx
@@ -1,7 +1,7 @@
-import { useEffect, useState } from "react";
+//@ts-nocheck
+import React, { useState } from "react";
 import Heading from "../ui/Heading";
 import Row from "../ui/Row";
-import { getCabins } from "../services/apiCabins";
 import CabinTable from "../features/cabins/CabinTable";
 import CreateCabinForm from "../features/cabins/CreateCabinForm";
 import Button from "../ui/Button";
